refactor, replace anchor tags with Link components for navigation consistency
