+++ src/App.tsx
@@ -1,6 +1,7 @@
 import { useState } from 'react'
 import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
+import { Link } from 'react-router'
 
 function App() {
   const [count, setCount] = useState(0)
@@ -8,12 +9,12 @@ function App() {
   return (
     <>
       <div>
-        <a href="https://vite.dev" target="_blank">
+        <Link to="https://vite.dev" target="_blank">
           <img src={viteLogo} alt="Vite logo" />
-        </a>
-        <a href="https://react.dev" target="_blank">
+        </Link>
+        <Link to="https://react.dev" target="_blank">
           <img src={reactLogo} alt="React logo" />
-        </a>
+        </Link>
       </div>
       <h1>Vite + React</h1>
       <div>
